Updated for Client Side Routing - probably still has some errors
