+++ package.json
@@ -12,7 +12,9 @@
     "json-server": "^0.9.4",
     "mongoose": "^4.8.3",
     "react": "^15.3.2",
-    "react-dom": "^15.3.2"
+    "react-dom": "^15.3.2",
+    "react-router": "^4.0.0-beta.6",
+    "react-router-dom": "^4.0.0-beta.8"
   },
   "scripts": {
     "start": "nf start",
